% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyD3.R
\docType{package}
\name{shinyD3}
\alias{shinyD3}
\title{Launch shinyD3 with default browser}
\usage{
shinyD3()
}
\value{
a shiny app launched in the web/mobile page
}
\description{
Launch shinyD3 with default browser
}
\examples{
if(interactive()){
shinyD3()
}

}
