% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyD3.R
\docType{data}
\name{valid_examples}
\alias{valid_examples}
\title{Valid Examples}
\format{
An object of class \code{character} of length 1.
}
\usage{
valid_examples
}
\description{
Valid Examples
}
\keyword{datasets}
