% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyD3.R
\docType{package}
\name{shinyD3}
\alias{shinyD3}
\title{Launch shinyD3 with default browser}
\usage{
shinyD3(example)
}
\arguments{
\item{example}{A  a valid string example of visualization}
}
\value{
a shiny app with D3 library launched in the web page
}
\description{
Launch shinyD3 with default browser
}
\examples{
if(interactive()){
shinyD3(example= "barplot")
}

}
